refactoring, removed unnecessary
