+++ src/chatwindow/Main.java
@@ -10,7 +10,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("MainWindow.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("MainChatWindow.fxml"));
         primaryStage.setTitle("Online Chat");
         primaryStage.setScene(new Scene(root));
         primaryStage.show();
