+++ src/chatwindow/MainChatWindow.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.text.Font?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="350.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="chatwindow.MainWindowController">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="350.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="chatwindow.MainChatWindowController">
    <top>
       <MenuBar BorderPane.alignment="CENTER">
         <menus>
@@ -56,11 +56,11 @@
                   </BorderPane>
                </bottom>
                <center>
-                  <TableView fx:id="chatMessages" focusTraversable="false" BorderPane.alignment="CENTER">
+                  <TableView fx:id="chatMessagesTable" focusTraversable="false" BorderPane.alignment="CENTER">
                     <columns>
-                        <TableColumn fx:id="timeTableView" editable="false" maxWidth="66.0" minWidth="66.0" prefWidth="66.0" sortable="false" text="Time" />
-                      <TableColumn fx:id="userTableView" editable="false" minWidth="100.0" prefWidth="300.0" sortable="false" text="User" />
-                      <TableColumn fx:id="messageTableView" editable="false" maxWidth="1.7976931348623157E308" minWidth="-Infinity" prefWidth="185.0" sortable="false" text="Message" />
+                        <TableColumn fx:id="timeTableField" editable="false" maxWidth="66.0" minWidth="66.0" prefWidth="66.0" sortable="false" text="Time" />
+                      <TableColumn fx:id="userTableField" editable="false" minWidth="100.0" prefWidth="300.0" sortable="false" text="User" />
+                      <TableColumn fx:id="messageTableField" editable="false" maxWidth="1.7976931348623157E308" minWidth="-Infinity" prefWidth="185.0" sortable="false" text="Message" />
                     </columns>
                      <columnResizePolicy>
                         <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
