+++ src/chatwindow/MainChatWindowController.java
@@ -1,0 +1,93 @@
+package chatwindow;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
+import javafx.scene.control.cell.PropertyValueFactory;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+
+public class MainChatWindowController {
+
+    @FXML
+    private ListView<String> chatUsersList;
+
+    @FXML
+    private TableView<RowChatMessage> chatMessagesTable;
+
+    @FXML
+    private TableColumn<RowChatMessage, String> userTableField;
+
+    @FXML
+    private  TableColumn<RowChatMessage, String> messageTableField;
+
+    @FXML
+    private TableColumn<RowChatMessage, String> timeTableField;
+
+
+    @FXML
+    private Button sendMessageButton;
+
+    @FXML
+    private TextField sendMessageText;
+
+
+    private String currentUser = "Иванов Иван"; //В начальной реализации это пока что фиксированное значение
+
+    @FXML
+    void initialize() {
+        userTableField.setCellValueFactory(new PropertyValueFactory<>("user"));
+        messageTableField.setCellValueFactory(new PropertyValueFactory<>("message"));
+        timeTableField.setCellValueFactory(new PropertyValueFactory<>("time"));
+
+        ObservableList<String> chatUsers = FXCollections.observableArrayList("Петров Петр", "Федор Михайлович");
+        chatUsersList.setItems(chatUsers); //добавим несколько записей в поле с активными пользователями чата, для теста
+        chatUsersList.getItems().add(currentUser); //Добавляем "текущего" пользователя. Пока только для теста
+
+    }
+
+
+    @FXML
+    void sendMessage() {
+        sendMessageText.requestFocus(); //при вызове метода фокус сразу возвращается на sendMessageText
+
+        String message=sendMessageText.getText().trim(); //введенное сообщение
+
+        if (!message.isBlank()) { //если что-то введено, то добавляем сообщение
+            Date date = new Date(); //текущая дата и время
+            SimpleDateFormat timeFormat = new SimpleDateFormat("HH:mm:ss");
+            String currentTime = timeFormat.format(date); //Преобразуем время в нужный формат
+            chatMessagesTable.getItems().add(new RowChatMessage(currentTime, currentUser, message));
+            sendMessageText.clear();
+            sendMessageButton.setDisable(true); //после отправки сделаем кнопку неактивной. Кнопка станет активной, если что-то введено
+            int messagesCount = chatMessagesTable.getItems().size();
+            chatMessagesTable.scrollTo(messagesCount -1 ); //прокрутим к последнему сообшению
+        }
+    }
+
+
+    @FXML
+    void setActiveSendButton() { //метод вызывается при вводе текста
+        //если пусто, то кнопка неактивна
+        sendMessageButton.setDisable(sendMessageText.getText().trim().isBlank()); //если что-то введено, то кнопку делаем активной
+    }
+
+    @FXML
+    void exit() {
+        System.exit(0);
+    }
+
+    @FXML
+    void about() {
+        Alert about = new Alert(Alert.AlertType.INFORMATION);
+        about.setTitle("О программе");
+        about.setHeaderText("Online - чат");
+        about.setContentText("Курс Java Core. Продвинутый уровень.");
+        about.show();
+    }
+
+
+}
